--- src/main/java/com/tta/geumgiri/domain/Admin.java
@@ -1,4 +1,0 @@
-package com.tta.geumgiri.domain;
-
-public class Admin {
-}
