--- src/main/java/com/tta/geumgiri/domain/Mission.java
@@ -1,4 +1,0 @@
-package com.tta.geumgiri.domain;
-
-public class Mission {
-}
