+++ src/main/java/com/tta/geumgiri/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package com.tta.geumgiri.repository;
+package com.tta.geumgiri.member.repository;
 
-import com.tta.geumgiri.domain.Member;
+import com.tta.geumgiri.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
