+++ src/test/java/com/tta/geumgiri/member/controller/MemberApiControllerTest.java
@@ -1,0 +1,4 @@
+package com.tta.geumgiri.member.controller;
+
+public class MemberApiControllerTest {
+}
